+++ src/script/components/Avatar-m/GroupAvatar.tsx
