+++ src/script/components/Avatar-m/PlaceholderAvatar/DefaultAvatarImage.tsx
