+++ src/script/components/Avatar-m/TemporaryGuestAvatar.tsx
