+++ src/script/components/InputBar/components/TypingIndicator/TypingIndicator.tsx
@@ -19,7 +19,7 @@
 
 import {FC} from 'react';
 
-import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
+import {Avatar, AVATAR_SIZE} from 'Components/Avatar-m';
 import {Icon} from 'Components/Icon';
 import {t} from 'Util/LocalizerUtil';
