+++ src/script/components/calling/CallParticipantsListItem/CallParticipantsListItem.tsx
@@ -21,7 +21,7 @@ import React from 'react';
 
 import {TabIndex} from '@wireapp/react-ui-kit/lib/types/enums';
 
-import {Avatar, AVATAR_SIZE} from 'Components/Avatar';
+import {Avatar, AVATAR_SIZE} from 'Components/Avatar-m';
 import {callParticipantListWrapper} from 'Components/calling/CallParticipantsListItem/CallParticipantsListItem.styles';
 import {ParticipantItemContent} from 'Components/ParticipantItemContent';
 import {listItem} from 'Components/ParticipantItemContent/ParticipantItem.styles';
